scripts/requirements.txt
kagglehub>=0.2.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tensorflow>=2.15.0
flask>=3.0.0
pyserial>=3.5
